requirements.txt
streamlit
python-docx
PyPDF2
openai
chromadb
sentence-transformers
